Propose removal of AuditIngestionComponent (replace with ctor in AuditIngestor)
